general update

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: spacotto <spacotto@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/14 12:14:30 by spacotto          #+#    #+#             */
-/*   Updated: 2025/10/14 12:18:55 by spacotto         ###   ########.fr       */
+/*   Updated: 2025/10/14 16:02:06 by spacotto         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
